requirements.txt
numpy==1.17.0
Pillow==6.1.0
ProgressPrinter==0.1.1
